펌뱅킹 요청

+++ castle-banking-service/src/main/java/com/castle/wookpay/banking/domain/command/RequestFirmBankingCommand.java
@@ -8,7 +8,7 @@ public record RequestFirmBankingCommand(
 		String fromBankAccountNumber,
 		String toBankName,
 		String toBankAccountNumber,
-		int moneyAmount
+		Long moneyAmount
 ) {
 
 }
