+++ castle-banking-service/src/main/java/com/castle/wookpay/banking/domain/result/RequestFirmBankingResult.java
@@ -8,7 +8,6 @@ public record RequestFirmBankingResult(
 		String fromBankAccountNumber,
 		String toBankName,
 		String toBankAccountNumber,
-		int moneyAmount
+		Long moneyAmount
 ) {
-
 }
